pyproject.toml
[project]
name = "timcam"
dynamic = ["version"]

dependencies = [
    "ezdxf >= 1.3.4",
    "pyvoronoi >= 1.1.5",
    "pillow >= 11.0.0",
    "pycairo",
    "keke",
    "toposort",
    "pyclipper",
    "vmodule",
]
requires-python = ">= 3.10"

# optional-depends
# test=
# pytest
# pytest-cov

